+++ kat/src/main/java/plus/kat/anno/Expose.java
@@ -54,7 +54,7 @@ public @interface Expose {
      * @see Flag
      * @since 0.0.4
      */
-    int require() default 0;
+    long require() default 0;
 
     /**
      * Returns the specified {@link Coder} or the pointing {@link Class}
